+++ app/src/test/java/org/simple/clinic/home/overdue/search/OverdueSearchEffectHandlerTest.kt
@@ -302,4 +302,20 @@ class OverdueSearchEffectHandlerTest {
     verifyNoMoreInteractions(uiActions)
     effectHandlerTestCase.assertNoOutgoingEvents()
   }
+
+  @Test
+  fun `when select all appointment ids effect is received, then add all selected ids`() {
+    // given
+    val allAppointmentIds = setOf(
+        UUID.fromString("9737ad9e-70f6-454e-9226-6eaa01c62645"))
+
+    // when
+    effectHandlerTestCase.dispatch(SelectAllAppointmentIds(allAppointmentIds))
+
+    // then
+    verifyZeroInteractions(uiActions)
+    verify(overdueAppointmentSelector).addSelectedIds(allAppointmentIds)
+    verifyNoMoreInteractions(overdueAppointmentSelector)
+    effectHandlerTestCase.assertNoOutgoingEvents()
+  }
 }